// src/app/feature-modules/main/sidebarconfigure/file-triggers/data/constant-value.enum.ts
export enum Constant {
    CONTROLLER = "/SymFileTriggers",
    PATH = "/main/configure/file-triggers",
    UNCLICKABLE = "unclickable",
    CLICKABLE = "clickable",
    CREATE = "/create",
    EDIT = "/edit",
    DELETE = "/delete",
    SORT_BY = "trigger_id",
    SORT_ORDER = "asc",
    FILTER_BY = "id",
    SAVE = "Save",
    UPDATE = "Update"
}